Rename Person package to Task package
