+++ src/main/java/seedu/geekeep/logic/commands/AddCommand.java
@@ -5,14 +5,14 @@ import java.util.Set;
 
 import seedu.geekeep.commons.exceptions.IllegalValueException;
 import seedu.geekeep.logic.commands.exceptions.CommandException;
-import seedu.geekeep.model.person.Address;
-import seedu.geekeep.model.person.Email;
-import seedu.geekeep.model.person.Name;
-import seedu.geekeep.model.person.Person;
-import seedu.geekeep.model.person.Phone;
-import seedu.geekeep.model.person.UniquePersonList;
 import seedu.geekeep.model.tag.Tag;
 import seedu.geekeep.model.tag.UniqueTagList;
+import seedu.geekeep.model.task.Address;
+import seedu.geekeep.model.task.Email;
+import seedu.geekeep.model.task.Name;
+import seedu.geekeep.model.task.Person;
+import seedu.geekeep.model.task.Phone;
+import seedu.geekeep.model.task.UniquePersonList;
 
 /**
  * Adds a person to the address book.
