+++ src/main/java/seedu/geekeep/logic/parser/ParserUtil.java
@@ -13,12 +13,12 @@ import java.util.stream.Collectors;
 
 import seedu.geekeep.commons.exceptions.IllegalValueException;
 import seedu.geekeep.commons.util.StringUtil;
-import seedu.geekeep.model.person.Address;
-import seedu.geekeep.model.person.Email;
-import seedu.geekeep.model.person.Name;
-import seedu.geekeep.model.person.Phone;
 import seedu.geekeep.model.tag.Tag;
 import seedu.geekeep.model.tag.UniqueTagList;
+import seedu.geekeep.model.task.Address;
+import seedu.geekeep.model.task.Email;
+import seedu.geekeep.model.task.Name;
+import seedu.geekeep.model.task.Phone;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes
