+++ src/main/java/seedu/geekeep/model/task/Person.java
@@ -1,4 +1,4 @@
-package seedu.geekeep.model.person;
+package seedu.geekeep.model.task;
 
 import java.util.Objects;
