+++ src/main/java/seedu/geekeep/ui/BrowserPanel.java
@@ -6,7 +6,7 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.Region;
 import javafx.scene.web.WebView;
 import seedu.geekeep.commons.util.FxViewUtil;
-import seedu.geekeep.model.person.ReadOnlyPerson;
+import seedu.geekeep.model.task.ReadOnlyPerson;
 
 /**
  * The Browser Panel of the App.
