+++ src/test/java/guitests/guihandles/PersonCardHandle.java
@@ -8,8 +8,8 @@ import javafx.scene.Node;
 import javafx.scene.control.Labeled;
 import javafx.scene.layout.Region;
 import javafx.stage.Stage;
-import seedu.geekeep.model.person.ReadOnlyPerson;
 import seedu.geekeep.model.tag.UniqueTagList;
+import seedu.geekeep.model.task.ReadOnlyPerson;
 
 /**
  * Provides a handle to a person card in the person list panel.
