+++ src/test/java/guitests/guihandles/PersonListPanelHandle.java
@@ -13,8 +13,8 @@ import javafx.scene.Node;
 import javafx.scene.control.ListView;
 import javafx.stage.Stage;
 import seedu.geekeep.TestApp;
-import seedu.geekeep.model.person.Person;
-import seedu.geekeep.model.person.ReadOnlyPerson;
+import seedu.geekeep.model.task.Person;
+import seedu.geekeep.model.task.ReadOnlyPerson;
 import seedu.geekeep.testutil.TestUtil;
 
 /**
