+++ src/test/java/seedu/geekeep/model/person/NameTest.java
@@ -5,6 +5,8 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
+import seedu.geekeep.model.task.Name;
+
 public class NameTest {
 
     @Test
