+++ src/test/java/seedu/geekeep/testutil/TestUtil.java
@@ -30,14 +30,14 @@ import seedu.geekeep.commons.exceptions.IllegalValueException;
 import seedu.geekeep.commons.util.FileUtil;
 import seedu.geekeep.commons.util.XmlUtil;
 import seedu.geekeep.model.TaskManager;
-import seedu.geekeep.model.person.Address;
-import seedu.geekeep.model.person.Email;
-import seedu.geekeep.model.person.Name;
-import seedu.geekeep.model.person.Person;
-import seedu.geekeep.model.person.Phone;
-import seedu.geekeep.model.person.ReadOnlyPerson;
 import seedu.geekeep.model.tag.Tag;
 import seedu.geekeep.model.tag.UniqueTagList;
+import seedu.geekeep.model.task.Address;
+import seedu.geekeep.model.task.Email;
+import seedu.geekeep.model.task.Name;
+import seedu.geekeep.model.task.Person;
+import seedu.geekeep.model.task.Phone;
+import seedu.geekeep.model.task.ReadOnlyPerson;
 import seedu.geekeep.storage.XmlSerializableAddressBook;
 
 /**
