+++ src/test/java/seedu/geekeep/testutil/TypicalTestPersons.java
@@ -2,8 +2,8 @@ package seedu.geekeep.testutil;
 
 import seedu.geekeep.commons.exceptions.IllegalValueException;
 import seedu.geekeep.model.TaskManager;
-import seedu.geekeep.model.person.Person;
-import seedu.geekeep.model.person.UniquePersonList;
+import seedu.geekeep.model.task.Person;
+import seedu.geekeep.model.task.UniquePersonList;
 
 /**
  *
